+++ src/pico8/api/param.rs
@@ -41,6 +41,7 @@ pub struct Pico8<'w, 's> {
     pub(crate) sprite_sheets: ResMut<'w, Assets<SpriteSheet>>,
     pub(crate) p8_maps: ResMut<'w, Assets<P8Map>>,
     pub(crate) audio_banks: ResMut<'w, Assets<AudioBank>>,
+    pub(crate) cart_args: ResMut<'w, CartArgs>,
 }
 
 impl Pico8<'_, '_> {